moved get_sector_for_point into UniformGrid where it makes a little more sense
